适配group编号修改；代码清理

+++ configs/src/data_server_configs.rs
@@ -7,7 +7,7 @@ pub struct DataServerConfigs {
     // 数据根目录
     pub root_dir_path: String,
     // 文件最大大小
-    pub max_file_size: u128,
+    pub max_file_size: u64,
     // 文件集最大数量
     pub max_set_size: u32,
     // 文件序列最大数量
